property creation

+++ Application/Validators/Property/CreatePropertyDtoValidator.cs
@@ -18,7 +18,7 @@ namespace Application.Validators.Property
                 .NotEmpty().WithMessage("Building number is required.")
                 .MaximumLength(20).WithMessage("Building number is too long.");
 
-            RuleFor(p => (int)p.PropertyType)
+            RuleFor(p => p.PropertyType)
                 .IsInEnum().WithMessage("Invalid Property Type.");
         }
     }
